Give renderShine an offset instead of a matrix bracket
Four callers drew a glow by pushing a matrix, translating to the spot and
popping again. A laser lights every fourth pixel of its beam and a light
barrier lit every one of them, so one beam across the screen spent hundreds
of matrix brackets on a sprite that is nothing but a soft disc: three GL calls
each, for a transform Engine::renderSprite was about to apply anyway.
renderShine takes the offset now and adds it to the sprite position, and the
default argument leaves the thirteen callers that glow on the spot alone.

That needs the sprite position to be a Vec2d. The laser and the light barrier
place their glows on half pixels, which the old Vec2i would have truncated;
every other call site hands it an integer vector, which converts exactly, and
both paths inside renderSprite had already promoted to double.

Measured against the frame HEAD draws, on a night-vision level holding a laser
and two light barriers: byte-identical at 640x480, and byte-identical again
under -nobatch, which is the arm that puts a fractional position through the
immediate glBegin path. What makes that comparison mean anything is the
control - two runs of one binary agree byte for byte, with the four random()
entry points returning the middle of their range and the level frozen at a
fixed tick. Without the int overload neutralised as well, the night vision's
own noise moved 65% of the pixels between two runs of the same binary.

The light barrier now lights every fourth pixel of its beam, the same as the
laser. That one is a visible change and deliberately so: 23685 of 307200
pixels move, and the mean brightness of the play area falls from 9.47 to 8.99.

+++ Blocks5/src/engine.cpp
@@ -2921,7 +2921,7 @@ bool Engine::screenshot()
 #endif
 }
 
-void Engine::renderSprite(const Vec2i& position,
+void Engine::renderSprite(const Vec2d& position,
 						  const Vec2i& positionOnTexture,
 						  const Vec2i& size,
 						  const Vec4d& color,
@@ -3072,7 +3072,7 @@ void Engine::endSpriteBatch()
 	spriteBatchOpen = false;
 }
 
-void Engine::queueSprite(const Vec2i& position,
+void Engine::queueSprite(const Vec2d& position,
 						 const Vec2i& halfSize,
 						 const Vec2i& otherHalf,
 						 int u0,
@@ -3160,7 +3160,7 @@ void Engine::queueSprite(const Vec2i& position,
 }
 
 void Engine::renderSprite(Texture* p_sprite,
-						  const Vec2i& position,
+						  const Vec2d& position,
 						  const Vec2i& positionOnTexture,
 						  const Vec2i& size,
 						  const Vec4d& color,
